+++ .dumi/theme/builtins/ThemeEditorPage/demos/nft-card.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import type { ComponentDemo } from 'antd-token-previewer-web3';
 
 export const components: Record<string, string[]> = {
-  'NFT Card': ['Basic NFT Card', 'Dark NFT', 'With Wagmi'],
+  'NFT Card': ['Basic NFT Card'],
 };
 
 export const demos: Record<string, ComponentDemo[]> = {
@@ -11,20 +11,6 @@ export const demos: Record<string, ComponentDemo[]> = {
       demo: React.createElement(require('@ant-design/web3/nft-card/demos/basic').default),
       tokens: [],
       key: 'Basic NFT Card',
-    },
-  ],
-  'Dark NFT': [
-    {
-      demo: React.createElement(require('@ant-design/web3/nft-card/demos/dark-mode').default),
-      tokens: [],
-      key: 'Dark NFT',
-    },
-  ],
-  'With Wagmi': [
-    {
-      demo: React.createElement(require('@ant-design/web3/nft-card/demos/wagmi').default),
-      tokens: [],
-      key: 'With Wagmi',
     },
   ],
 };