+++ .dumirc.ts
@@ -58,9 +58,9 @@ export default defineConfig({
     {
       name: 'keywords',
       content:
-        'web3, Ant Design Web3, Ant Design, dapp, frontend, components,web3, admin , Ant Design, ant design, react, alibaba',
+        'Ethereum, Solana, Ton, Bitcoin, web3, Ant Design Web3, Ant Design, antd, dapp, frontend, components, react, ui kit',
     },
-    { name: 'description', content: 'A Components for Web3, Based on Ant Design,' },
+    { name: 'description', content: 'A Components for Web3, Based on Ant Design.' },
     {
       property: 'og:site_name',
       content: 'Ant Design Web3',
@@ -73,7 +73,7 @@ export default defineConfig({
     {
       property: 'og:description',
       content:
-        'Ant Design Web3 是一个基于 Ant Design 的 Web3 React 组件库，它提供了一系列的组件，可以帮助你快速构建去中心化应用（DApp）。',
+        'Ant Design Web3 is a Web3 React component library based on Ant Design. It offers a suite of components that can help you quickly build decentralized applications (DApps). Ant Design Web3 是一个基于 Ant Design 的 Web3 React 组件库，它提供了一系列的组件，可以帮助你快速构建去中心化应用（DApp）。',
     },
     {
       name: 'apple-mobile-web-app-capable',