+++ docs/guide/demos/try-it-out/ethereum.tsx
@@ -1,0 +1,58 @@
+import React from 'react';
+import type { ConnectModalProps } from '@ant-design/web3';
+import { ConnectButton, Connector } from '@ant-design/web3';
+import {
+  MetaMask,
+  OkxWallet,
+  TokenPocket,
+  WagmiWeb3ConfigProvider,
+  WalletConnect,
+} from '@ant-design/web3-wagmi';
+import type { ConfigProviderProps } from 'antd';
+
+type SizeType = ConfigProviderProps['componentSize'];
+
+interface Props {
+  mode: ConnectModalProps['mode'];
+  size: SizeType;
+  quickConnect: boolean;
+}
+
+const App: React.FC<Props> = ({ mode, size, quickConnect }) => {
+  return (
+    <WagmiWeb3ConfigProvider
+      eip6963={{
+        autoAddInjectedWallets: true,
+      }}
+      walletConnect={{
+        projectId: YOUR_WALLET_CONNECT_PROJECT_ID,
+      }}
+      ens
+      wallets={[
+        MetaMask(),
+        WalletConnect(),
+        TokenPocket({
+          group: 'Popular',
+        }),
+        OkxWallet(),
+      ]}
+    >
+      <Connector
+        modalProps={{
+          mode,
+        }}
+      >
+        <ConnectButton
+          type="primary"
+          style={{
+            width: 'auto',
+          }}
+          size={size}
+          quickConnect={quickConnect}
+        />
+      </Connector>
+    </WagmiWeb3ConfigProvider>
+  );
+};
+
+export default App;