+++ docs/guide/demos/try-it-out/sui.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import type { ConnectModalProps } from '@ant-design/web3';
+import { ConnectButton, Connector } from '@ant-design/web3';
+import { Suiet, SuiWallet, SuiWeb3ConfigProvider } from '@ant-design/web3-sui';
+import type { ConfigProviderProps } from 'antd';
+
+type SizeType = ConfigProviderProps['componentSize'];
+
+interface Props {
+  mode: ConnectModalProps['mode'];
+  size: SizeType;
+  quickConnect: boolean;
+}
+
+const App: React.FC<Props> = ({ mode, size, quickConnect }) => {
+  return (
+    <SuiWeb3ConfigProvider balance autoConnect wallets={[Suiet(), SuiWallet()]}>
+      <Connector
+        modalProps={{
+          mode,
+        }}
+      >
+        <ConnectButton
+          type="primary"
+          style={{
+            width: 'auto',
+          }}
+          size={size}
+          quickConnect={quickConnect}
+        />
+      </Connector>
+    </SuiWeb3ConfigProvider>
+  );
+};
+
+export default App;