+++ packages/common/src/types.ts
@@ -283,6 +283,7 @@ export interface RequiredLocale {
     guideTipLearnMoreLinkText: string;
     walletPanelPlugin: string;
     walletConnectSuccess: string;
+    walletListEmpty: string;
     getWalletTipsTitle: string;
     getWalletTipsDesc: string;
     linkWallet: string;